requirements.txt
fastrtc[vad]
twilio
openai
gradio
pydantic_settings
